Primeira Versao API

+++ src/Backend/ControleAcesso.Dominio/Entidades/Socio.cs
@@ -10,7 +10,7 @@ public class Socio : EntidadeBase
     
     public string? Senha { get; set; }
     
-    public int PlanoId { get; set; }
+    public long PlanoId { get; set; }
 
     public Plano Plano { get; set; }
 }
